+++ CHANGELOG.md
@@ -4,6 +4,9 @@ CHANGELOG for Sulu
 * dev-develop
     * BUGFIX      #3206 [SnippetBundle]       Corrected translations in copy locale and open ghost overlay
     * ENHANCEMENT #3206 [ContentBundle]       Made translations for copy locale and open ghost overlay changeable
+
+* 1.5.0-RC2 (2017-02-20)
+    * BUGFIX      #3216 [ContentBundle]       Added author and authored to reserved-property-names
     * BUGFIX      #3210 [MediaBundle]         Allow same image format key if formats are identical
     * BUGFIX      #3200 [SecurityBundle]      Fixed broken UserManager when used without Security
     * BUGFIX      #3183 [ContentBundle]       Fixed Windows xinclude error